further refactoring of the C code
