+++ tinywl/src/seat.c
@@ -1,0 +1,51 @@
+#include "seat.h"
+#include "tinywl.h"
+
+void seat_request_cursor(struct wl_listener *listener, void *data) {
+  struct tinywl_server *server =
+      wl_container_of(listener, server, request_cursor);
+  /* This event is raised by the seat when a client provides a cursor image */
+  struct wlr_seat_pointer_request_set_cursor_event *event = data;
+  struct wlr_seat_client *focused_client =
+      server->seat->pointer_state.focused_client;
+  /* This can be sent by any client, so we check to make sure this one is
+   * actually has pointer focus first. */
+  if (focused_client == event->seat_client) {
+    /* Once we've vetted the client, we can tell the cursor to use the
+     * provided surface as the cursor image. It will set the hardware cursor
+     * on the output that it's currently on and continue to do so as the
+     * cursor moves between outputs. */
+    wlr_cursor_set_surface(server->cursor, event->surface, event->hotspot_x,
+                           event->hotspot_y);
+  }
+}
+
+void seat_request_set_selection(struct wl_listener *listener, void *data) {
+  /* This event is raised by the seat when a client wants to set the selection,
+   * usually when the user copies something. wlroots allows compositors to
+   * ignore such requests if they so choose, but in tinywl we always honor
+   */
+  struct tinywl_server *server =
+      wl_container_of(listener, server, request_set_selection);
+  struct wlr_seat_request_set_selection_event *event = data;
+  wlr_seat_set_selection(server->seat, event->source, event->serial);
+}
+
+void reset_cursor_mode(struct tinywl_server *server) {
+  /* Reset the cursor mode to passthrough. */
+  server->cursor_mode = TINYWL_CURSOR_PASSTHROUGH;
+  server->grabbed_toplevel = NULL;
+}
+
+void initialize_seat(struct tinywl_server *server) {
+  wl_list_init(&server->keyboards);
+  server->new_input.notify = server_new_input;
+  wl_signal_add(&server->backend->events.new_input, &server->new_input);
+  server->seat = wlr_seat_create(server->wl_display, "seat0");
+  server->request_cursor.notify = seat_request_cursor;
+  wl_signal_add(&server->seat->events.request_set_cursor,
+                &server->request_cursor);
+  server->request_set_selection.notify = seat_request_set_selection;
+  wl_signal_add(&server->seat->events.request_set_selection,
+                &server->request_set_selection);
+}
