+++ tinywl/tinywl.c
@@ -48,273 +48,6 @@ bool cycle_windows(struct tinywl_server *server) {
   return true;
 }
 
-static void server_new_pointer(struct tinywl_server *server,
-                               struct wlr_input_device *device) {
-  /* We don't do anything special with pointers. All of our pointer handling
-   * is proxied through wlr_cursor. On another compositor, you might take this
-   * opportunity to do libinput configuration on the device to set
-   * acceleration, etc. */
-  wlr_cursor_attach_input_device(server->cursor, device);
-}
-
-static void server_new_input(struct wl_listener *listener, void *data) {
-  /* This event is raised by the backend when a new input device becomes
-   * available. */
-  struct tinywl_server *server = wl_container_of(listener, server, new_input);
-  struct wlr_input_device *device = data;
-  switch (device->type) {
-  case WLR_INPUT_DEVICE_KEYBOARD:
-    server_new_keyboard(server, device);
-    break;
-  case WLR_INPUT_DEVICE_POINTER:
-    server_new_pointer(server, device);
-    break;
-  default:
-    break;
-  }
-  /* We need to let the wlr_seat know what our capabilities are, which is
-   * communiciated to the client. In TinyWL we always have a cursor, even if
-   * there are no pointer devices, so we always include that capability. */
-  uint32_t caps = WL_SEAT_CAPABILITY_POINTER;
-  if (!wl_list_empty(&server->keyboards)) {
-    caps |= WL_SEAT_CAPABILITY_KEYBOARD;
-  }
-  wlr_seat_set_capabilities(server->seat, caps);
-}
-
-static void seat_request_cursor(struct wl_listener *listener, void *data) {
-  struct tinywl_server *server =
-      wl_container_of(listener, server, request_cursor);
-  /* This event is raised by the seat when a client provides a cursor image */
-  struct wlr_seat_pointer_request_set_cursor_event *event = data;
-  struct wlr_seat_client *focused_client =
-      server->seat->pointer_state.focused_client;
-  /* This can be sent by any client, so we check to make sure this one is
-   * actually has pointer focus first. */
-  if (focused_client == event->seat_client) {
-    /* Once we've vetted the client, we can tell the cursor to use the
-     * provided surface as the cursor image. It will set the hardware cursor
-     * on the output that it's currently on and continue to do so as the
-     * cursor moves between outputs. */
-    wlr_cursor_set_surface(server->cursor, event->surface, event->hotspot_x,
-                           event->hotspot_y);
-  }
-}
-
-static void seat_request_set_selection(struct wl_listener *listener,
-                                       void *data) {
-  /* This event is raised by the seat when a client wants to set the selection,
-   * usually when the user copies something. wlroots allows compositors to
-   * ignore such requests if they so choose, but in tinywl we always honor
-   */
-  struct tinywl_server *server =
-      wl_container_of(listener, server, request_set_selection);
-  struct wlr_seat_request_set_selection_event *event = data;
-  wlr_seat_set_selection(server->seat, event->source, event->serial);
-}
-
-void reset_cursor_mode(struct tinywl_server *server) {
-  /* Reset the cursor mode to passthrough. */
-  server->cursor_mode = TINYWL_CURSOR_PASSTHROUGH;
-  server->grabbed_toplevel = NULL;
-}
-
-static void process_cursor_move(struct tinywl_server *server, uint32_t time) {
-  /* Move the grabbed toplevel to the new position. */
-  struct tinywl_toplevel *toplevel = server->grabbed_toplevel;
-  wlr_scene_node_set_position(&toplevel->scene_tree->node,
-                              server->cursor->x - server->grab_x,
-                              server->cursor->y - server->grab_y);
-}
-
-static void process_cursor_resize(struct tinywl_server *server, uint32_t time) {
-  struct tinywl_toplevel *toplevel = server->grabbed_toplevel;
-  const int border_thickness = 5;
-
-  double border_x = server->cursor->x - server->grab_x;
-  double border_y = server->cursor->y - server->grab_y;
-  int new_left = server->grab_geobox.x;
-  int new_right = server->grab_geobox.x + server->grab_geobox.width;
-  int new_top = server->grab_geobox.y;
-  int new_bottom = server->grab_geobox.y + server->grab_geobox.height;
-
-  if (server->resize_edges & WLR_EDGE_TOP) {
-    new_top = border_y;
-    if (new_top >= new_bottom - 2 * border_thickness) {
-      new_top = new_bottom - 2 * border_thickness - 1;
-    }
-  } else if (server->resize_edges & WLR_EDGE_BOTTOM) {
-    new_bottom = border_y;
-    if (new_bottom <= new_top + 2 * border_thickness) {
-      new_bottom = new_top + 2 * border_thickness + 1;
-    }
-  }
-  if (server->resize_edges & WLR_EDGE_LEFT) {
-    new_left = border_x;
-    if (new_left >= new_right - 2 * border_thickness) {
-      new_left = new_right - 2 * border_thickness - 1;
-    }
-  } else if (server->resize_edges & WLR_EDGE_RIGHT) {
-    new_right = border_x;
-    if (new_right <= new_left + 2 * border_thickness) {
-      new_right = new_left + 2 * border_thickness + 1;
-    }
-  }
-
-  struct wlr_box geo_box;
-  wlr_xdg_surface_get_geometry(toplevel->xdg_toplevel->base, &geo_box);
-
-  wlr_scene_node_set_position(&toplevel->scene_tree->node, new_left - geo_box.x,
-                              new_top - geo_box.y);
-
-  // Calculate new dimensions accounting for borders
-  int new_width = new_right - new_left - (2 * border_thickness);
-  int new_height = new_bottom - new_top - (2 * border_thickness);
-
-  // Ensure minimum size
-  if (new_width < 1)
-    new_width = 1;
-  if (new_height < 1)
-    new_height = 1;
-
-  wlr_xdg_toplevel_set_size(toplevel->xdg_toplevel, new_width, new_height);
-  update_border_position(toplevel);
-}
-
-static void process_cursor_motion(struct tinywl_server *server, uint32_t time) {
-  /* If the mode is non-passthrough, delegate to those functions. */
-  if (server->cursor_mode == TINYWL_CURSOR_MOVE) {
-    process_cursor_move(server, time);
-    return;
-  } else if (server->cursor_mode == TINYWL_CURSOR_RESIZE) {
-    process_cursor_resize(server, time);
-    return;
-  }
-
-  /* First check for layer surfaces */
-  double sx, sy;
-  struct wlr_seat *seat = server->seat;
-  struct wlr_surface *surface = NULL;
-
-  /* Check layer surfaces first */
-  struct wlr_scene_node *node =
-      wlr_scene_node_at(&server->scene->tree.node, server->cursor->x,
-                        server->cursor->y, &sx, &sy);
-  if (node && node->type == WLR_SCENE_NODE_BUFFER) {
-    struct wlr_scene_buffer *scene_buffer = wlr_scene_buffer_from_node(node);
-    struct wlr_scene_surface *scene_surface =
-        wlr_scene_surface_try_from_buffer(scene_buffer);
-    if (scene_surface) {
-      surface = scene_surface->surface;
-    }
-  }
-
-  /* If no layer surface found, check for regular toplevels */
-  if (!surface) {
-    struct tinywl_toplevel *toplevel = desktop_toplevel_at(
-        server, server->cursor->x, server->cursor->y, &surface, &sx, &sy);
-
-    if (!toplevel) {
-      /* If there's no toplevel, set the default cursor */
-      wlr_cursor_set_xcursor(server->cursor, server->cursor_mgr, "default");
-    }
-  }
-
-  if (surface) {
-    wlr_seat_pointer_notify_enter(seat, surface, sx, sy);
-    wlr_seat_pointer_notify_motion(seat, time, sx, sy);
-  } else {
-    wlr_seat_pointer_clear_focus(seat);
-  }
-}
-
-static void server_cursor_motion(struct wl_listener *listener, void *data) {
-  /* This event is forwarded by the cursor when a pointer emits a _relative_
-   * pointer motion event (i.e. a delta) */
-  struct tinywl_server *server =
-      wl_container_of(listener, server, cursor_motion);
-  struct wlr_pointer_motion_event *event = data;
-  /* The cursor doesn't move unless we tell it to. The cursor automatically
-   * handles constraining the motion to the output layout, as well as any
-   * special configuration applied for the specific input device which
-   * generated the event. You can pass NULL for the device if you want to move
-   * the cursor around without any input. */
-  wlr_cursor_move(server->cursor, &event->pointer->base, event->delta_x,
-                  event->delta_y);
-  process_cursor_motion(server, event->time_msec);
-}
-
-static void server_cursor_motion_absolute(struct wl_listener *listener,
-                                          void *data) {
-  /* This event is forwarded by the cursor when a pointer emits an _absolute_
-   * motion event, from 0..1 on each axis. This happens, for example, when
-   * wlroots is running under a Wayland window rather than KMS+DRM, and you
-   * move the mouse over the window. You could enter the window from any edge,
-   * so we have to warp the mouse there. There is also some hardware which
-   * emits these events. */
-  struct tinywl_server *server =
-      wl_container_of(listener, server, cursor_motion_absolute);
-  struct wlr_pointer_motion_absolute_event *event = data;
-  wlr_cursor_warp_absolute(server->cursor, &event->pointer->base, event->x,
-                           event->y);
-  process_cursor_motion(server, event->time_msec);
-}
-
-static void server_cursor_button(struct wl_listener *listener, void *data) {
-  struct tinywl_server *server =
-      wl_container_of(listener, server, cursor_button);
-  struct wlr_pointer_button_event *event = data;
-
-  // Check for layer surface first
-  double sx, sy;
-  struct wlr_surface *surface =
-      layer_surface_at(server, server->cursor->x, server->cursor->y, &sx, &sy);
-
-  if (surface) {
-    // If we found a layer surface, focus it and forward the click
-    focus_layer_surface(server, surface);
-    wlr_seat_pointer_notify_button(server->seat, event->time_msec,
-                                   event->button, event->state);
-    return;
-  }
-
-  // If no layer surface, check for regular windows
-  surface = NULL;
-  struct tinywl_toplevel *toplevel = desktop_toplevel_at(
-      server, server->cursor->x, server->cursor->y, &surface, &sx, &sy);
-
-  wlr_seat_pointer_notify_button(server->seat, event->time_msec, event->button,
-                                 event->state);
-
-  if (event->state == WLR_BUTTON_RELEASED) {
-    reset_cursor_mode(server);
-  } else {
-    focus_toplevel(toplevel, surface);
-  }
-}
-
-static void server_cursor_axis(struct wl_listener *listener, void *data) {
-  /* This event is forwarded by the cursor when a pointer emits an axis event,
-   * for example when you move the scroll wheel. */
-  struct tinywl_server *server = wl_container_of(listener, server, cursor_axis);
-  struct wlr_pointer_axis_event *event = data;
-  /* Notify the client with pointer focus of the axis event. */
-  wlr_seat_pointer_notify_axis(server->seat, event->time_msec,
-                               event->orientation, event->delta,
-                               event->delta_discrete, event->source);
-}
-
-static void server_cursor_frame(struct wl_listener *listener, void *data) {
-  /* This event is forwarded by the cursor when a pointer emits an frame
-   * event. Frame events are sent after regular pointer events to group
-   * multiple events together. For instance, two axis events may happen at the
-   * same time, in which case a frame event won't be sent in between. */
-  struct tinywl_server *server =
-      wl_container_of(listener, server, cursor_frame);
-  /* Notify the client with pointer focus of the frame event. */
-  wlr_seat_pointer_notify_frame(server->seat);
-}
-
 static void output_frame(struct wl_listener *listener, void *data) {
   struct tinywl_output *output = wl_container_of(listener, output, frame);
   struct wlr_scene *scene = output->server->scene;
@@ -468,23 +201,6 @@ void begin_interactive(struct tinywl_toplevel *toplevel,
     server->resize_edges = edges;
   }
 }
-
-// Function declarations
-char *parse_arguments(int argc, char *argv[]);
-void initialize_output_layout(struct tinywl_server *server);
-void initialize_scene(struct tinywl_server *server);
-void initialize_xdg_shell(struct tinywl_server *server);
-void initialize_cursor(struct tinywl_server *server);
-void initialize_seat(struct tinywl_server *server);
-const char *start_backend(struct tinywl_server *server);
-void run_startup_command(const char *startup_cmd);
-void cleanup(struct tinywl_server *server);
-struct tinywl_server *server_create();
-void server_destroy(struct tinywl_server *server);
-bool server_init(struct tinywl_server *server);
-const char *server_start(struct tinywl_server *server);
-void server_run(struct tinywl_server *server);
-void server_set_startup_command(const char *cmd);
 
 // New server functions
 struct tinywl_server *server_create() {
@@ -692,19 +408,6 @@ void initialize_cursor(struct tinywl_server *server) {
   wl_signal_add(&server->cursor->events.axis, &server->cursor_axis);
   server->cursor_frame.notify = server_cursor_frame;
   wl_signal_add(&server->cursor->events.frame, &server->cursor_frame);
-}
-
-void initialize_seat(struct tinywl_server *server) {
-  wl_list_init(&server->keyboards);
-  server->new_input.notify = server_new_input;
-  wl_signal_add(&server->backend->events.new_input, &server->new_input);
-  server->seat = wlr_seat_create(server->wl_display, "seat0");
-  server->request_cursor.notify = seat_request_cursor;
-  wl_signal_add(&server->seat->events.request_set_cursor,
-                &server->request_cursor);
-  server->request_set_selection.notify = seat_request_set_selection;
-  wl_signal_add(&server->seat->events.request_set_selection,
-                &server->request_set_selection);
 }
 
 bool initialize_backend_renderer_allocator(struct tinywl_server *server) {
